Feature/add hacking minigame to my frontend (#53)
* [MY] Remove remnants of the old OCS port. That was promoted to the MZ frontend

* [MY] Fix compilation on platforms

* [MY] Add first draft of the Hacking Mini-game

It is optional and text only for now. Only PC DOS and Master System will support it for now. Amstrad CPC will most likely as well.

* [MY] Revamp clean target for Speccy

* [MY] Add hacking mini game to ia16

Forgot to include this in the previous commits

* [MY] properly implement clearScreen for ia16

Forgot to change the implementation to differentiate between 386 and ia16

* [MY] Replace clearScreen with a forward call to init on ia16

I hate this, but seems like we're out of space in the segment...

* [MY] Disable hacking minigame for ia16

Still too big - there might be not enough memory for it...yet. But I'm not giving up just yet.

* [MY] forgot to remove one extra define from ia16 makefile

Co-authored-by: Daniel Monteiro <monteiroquiet@gmail.com>