Feature/add opengl port (#68)
* [GL] add initial version, based on the MX frontend. No GL code is present yet

This also removes the platforms not supported by this frontend.

* [GL] Get basic functionality in place, while also moving to better named folders

For now, it's mostly solid 2D shapes, but proper 3D is just around the corner.

* [GL] rename target to Derelict64

* [GL] (partially) fix bitmap loading

I can see text and..sometimes, some images, in the wrong colours. It's getting there!

* [GL] Mostly fix the 2D part, making the menus are all navigable

It's almost impossible to tell the GL version from the raster version

* [GL] add some basic support for the 3D levels

* [GL] Add support for SDL2 version...well, mostly

Font rendering is still faulty, but everything else is on par with the Cocoa version.

* [GL] rename CMake target for GL frontend

* [GL] add missing geometry types

Rendering still looks wonky, but at least there are a lot less holes.

* [MX] Rename Xcode target name to Derelict32

* [GL] Move the OpenGL view to the Xcode2 dir

* [GL] Fix global mismatch

* [GL] rename SDL2 Makefile

* [GL] Add *almost* functional N64 Makefile

* [GL] Fix asset bundling on N64

* [GL] Fix global types

* [GL] N64 version is capable of reaching most of the main menu

but leaks memory badly. But it moves!

* [GL] ensure the font texture is uploaded before trying to write text

* [GL] fix memory leak with the expanded texture when submitting to the GPU

* [GL] copy updated data file to the N64 filesystem when assembling a new version

* [GL] fix mistake when evaluating power of 2 textures before uploading to video memory

* [GL] fix texture loading on N64

Turns out I was memsetting the wrong buffer and only loading textures with 0s.

* [GL] fix texture upload on the N64

By making all textures uploaded as 16 bits, but using special flags and such for the N64

* [GL] fix trying to render the dropshadow before uploading the font to GPU

Since the dropshadow uses a char from the font.

* [GL] Fix font rendering distortion

* [GL] Restore Main Menu to the default, along with some fixes

- fix memory leak with the second logo bitmap
- remove Quit option from the N64 port

* [MX] refactor: replaces naked screen resolution numbers with contants

* [MX] fix: global type mismatch

* [GL] supports rendering the level 3D geometry on the N64!

A lot of the cruft from TMR was removed, as well as reducing some textures to 16x16. The size is no longer fixed, meaning we can have textures on the size we might want.

* [GL] Add initial support for TTF fonts on the N64

* [GL] Better handle text layouting, including line breaks

* [GL] Add the basis for the in-game UI drawing

* [GL] enable filtering to the in game textures only

* [GL] Fix tesselation on the 3D level shapes

* [GL] Fix geometry precision issues that caused the dancing cube bug

* [GL] Fix settings for Xcode 4

* [GL] add SDL1 makefile, along with fixes to make it work

Sadly, it doesn't support Emscripten (yet?).

* [GL] show items on the map

Some items have issues on their billboards, but it mostly work.

* [GL] fix crashes when trying to interact with the game objects

* [GL] Add graphical output to log messages

* [GL] Make the game (almost) playable on the N64 controller

A key for switching the current item is still pending

* [GL] Fix playability on the N64 controller

* [GL] fix the logic inversion bug for the inclusion of the Quit option between N64 and desktop

* [GL] update assets?

* [GL] fix room transition animation...mostly

* [GL] fix texture rotations

* [GL] fix billboard rendering

* [GL] Remove the old "loading" popup

It doesn't make sense anymore

* [GL] Fix game visuals and aesthetics

Had to widen the visilibility algorithm, but it did work fine and the game now plays in fullscreen, with the textures rotating nicely and properly aligned.

* [GL] display the splash screen images by tiling it

Thanks ImageMagick! :D

* [GL] Display map during gameplay

* [GL] visuals on hacking minigame

* [GL] restore belle and stdmatts textures

Erased it by mistake

* [GL] Fix remaining glitches on 2D screens

The game is now...feature complete?

* [GL] Make the Xcode version build well on Xcode2

* [GL] Ignore transient Xcode files

* [MY] Fix typo in CMakeLists

* [GL] fix SDL2 port

Out of sheer luck, it was running fine on MacOS, but not on Linux. This flag fixes everything.