[MX] Reduce static storage memory footprint
Moving a lot of the resources into dynamically allocated storage that can be potentially switched in and out of memory during runtime.
Right now, it leaks memory like crazy.

Thankfully, a lot of the flow was also simplificed, reducing a lot of the memory copying around and taking more advantage of globals