[MX] Fix AGS compilation
Turns out I had to include a pair of const to the variable declarations inside the generated basepfs.h. Even more surprising is that if I go back to the merge that added AGS support, it wouldn't cleanly link without that const either. I must find a way to automate this.